Remove prisma.ts, use a single database across test files

+++ cil-api/tests/setup.ts
@@ -1,8 +1,8 @@
 import { v4 as uuidv4 } from 'uuid';
-import { Context, grpc, proto } from 'cil-lib';
+import { Context, grpc, PrismaClient, proto } from 'cil-lib';
 import { OnboardingClient } from 'cil-lib/dist/main/lib/protos';
 import { OnboardingServer } from '../src/lib/api';
-import { prisma } from 'cil-lib/src/lib/utils/prisma';
+
 const { execSync } = require('child_process');
 const path = require('path');
 
@@ -28,44 +28,65 @@ export function prismaTestContext() {
     'schema.prisma'
   );
 
-  let dbName;
   return {
     async before() {
-      await prisma.$disconnect();
+      // store global prisma instance
+      await global.prisma?.$disconnect();
 
-      const splitDatabase = dbUrl.split('/');
-      const withoutDatabase = splitDatabase.splice(0, splitDatabase.length - 1);
-      const withDatabase = withoutDatabase;
+      // global variable to check if a test database has been created
+      if (global.isTestDbCreated == undefined) {
+        // global database name
+        global.testDb = `test-${uuidv4()}`;
+        // global counter that keeps track of how many test files are running
+        global.counter = 1;
 
-      // Generate a unique database for this test context
-      dbName = `test-${uuidv4()}`;
-      withDatabase.push(dbName);
+        const splitDatabase = dbUrl.split('/');
+        const withoutDatabase = splitDatabase.splice(
+          0,
+          splitDatabase.length - 1
+        );
+        const withDatabase = withoutDatabase;
 
-      const databaseUrl = withDatabase.join('/');
-      global.process.env.DATABASE_URL = `${databaseUrl}?schema=public`;
+        withDatabase.push(global.testDb);
 
-      // Run the migrations to ensure our schema has the required structure
-      execSync(`${PATH} ${prismaBinary} db push --schema=${schemaFilePath}`, {
-        env: { DATABASE_URL: global.process.env.DATABASE_URL },
-      });
+        const databaseUrl = withDatabase.join('/');
+        global.process.env.DATABASE_URL = `${databaseUrl}?schema=public`;
 
-      // Get prisma connected to the generated database
+        // Run the migrations to ensure our schema has the required structure
+        execSync(`${PATH} ${prismaBinary} db push --schema=${schemaFilePath}`, {
+          env: { DATABASE_URL: global.process.env.DATABASE_URL },
+        });
+        global.isTestDbCreated = true;
+        // Get prisma connected to the generated database
+        global.prisma = new PrismaClient();
+        await global.prisma.$connect();
+      } else {
+        // Increase the counter to take account for the current test file
+        global.counter += 1;
+      }
 
-      await prisma.$connect();
-      return prisma;
+      return global.prisma;
     },
     async after() {
-      // Get list of tables
-      const res: [{ table_name: string }] =
-        await prisma.$queryRaw`SELECT table_name from information_schema.tables WHERE table_catalog = ${dbName} AND table_schema = ${'public'}`;
+      // Check if the test file is the last one. If true then clean up the database
+      if (global.counter == 1) {
+        // Get list of tables
+        const res: [{ table_name: string }] = await global.prisma
+          .$queryRaw`SELECT table_name from information_schema.tables WHERE table_catalog = ${
+          global.testDb
+        } AND table_schema = ${'public'}`;
 
-      // Drop the tables after the tests have completed
-      for (let r of res) {
-        const query = `DROP TABLE ${r.table_name} CASCADE`;
-        await prisma.$executeRawUnsafe(query);
+        // Drop the tables after the tests have completed
+        for (let r of res) {
+          const query = `DROP TABLE ${r.table_name} CASCADE`;
+          await global.prisma.$executeRawUnsafe(query);
+        }
+        // Release the Prisma Client connection
+        await global.prisma.$disconnect();
       }
-      // Release the Prisma Client connection
-      await prisma.$disconnect();
+
+      // Decrease counter to mark the test file has already finished
+      global.counter -= 1;
     },
   };
 }
